+++ src/scdice_metrics/utils/_diffusion_nn.py
@@ -5,7 +5,7 @@ import numpy as np
 import scipy
 from scipy.sparse import csr_matrix, issparse
 
-from scdcb_metrics import nearest_neighbors
+from scdice_metrics import nearest_neighbors
 
 logger = logging.getLogger(__name__)